+++ src/Map.php
@@ -276,11 +276,13 @@ final class Map implements \Countable
     /**
      * Reduce the map to a single value
      *
+     * @template I
      * @template R
-     * @param R $carry
-     * @param callable(R, T, S): R $reducer
      *
-     * @return R
+     * @param I $carry
+     * @param callable(I|R, T, S): R $reducer
+     *
+     * @return I|R
      */
     public function reduce($carry, callable $reducer)
     {