+++ src/Map/Primitive.php
@@ -277,7 +277,6 @@ final class Primitive implements Implementation
      */
     public function merge(Implementation $map): Implementation
     {
-        /** @psalm-suppress ArgumentTypeCoercion */
         return $map->reduce(
             $this,
             static fn(Implementation $carry, $key, $value): Implementation => ($carry)($key, $value),
@@ -309,11 +308,13 @@ final class Primitive implements Implementation
     }
 
     /**
+     * @template I
      * @template R
-     * @param R $carry
-     * @param callable(R, T, S): R $reducer
      *
-     * @return R
+     * @param I $carry
+     * @param callable(I|R, T, S): R $reducer
+     *
+     * @return I|R
      */
     public function reduce($carry, callable $reducer)
     {